make floating point rust version

+++ calc.rs
@@ -1,15 +1,15 @@
 use std::env;
 
 fn main() {
-    let mut stack: Vec<i32> = Vec::new();
+    let mut stack: Vec<f64> = Vec::new();
 
     // Loop through each argument provided
     for arg in env::args().skip(1) {
-        // Try to parse the argument as an integer
-        match arg.parse::<i32>() {
-            // If it's an integer, push it onto the stack
+        // Try to parse the argument as a floating point number
+        match arg.parse::<f64>() {
+            // If it's a number, push it onto the stack
             Ok(num) => stack.push(num),
-            // If it's not an integer, assume it's an operator
+            // If it's not a number, assume it's an operator
             Err(_) => {
                 // Pop the top two values off the stack
                 let b = stack.pop().expect("Stack underflow");
